fix type issue for remainingWorkerDay

+++ Entities/Concrete/Project.cs
@@ -24,7 +24,7 @@ public class Project : IEntity
     public int? WorkerDay { get; set; }
     public int? WorkerHour { get; set; }
     public int? RemainingContractBudget { get; set; }
-    public int? RemainingWorkerHour { get; set; }
+    public double? RemainingWorkerHour { get; set; }
     public EnumIsDeleted? IsDeleted { get; set; }
     public DateTime? CreatedAt { get; set; }
     public DateTime? ModifiedAt { get; set; }
